+++ example/lib/main.dart
@@ -1,8 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:sp_settings/settings_form.dart';
+import 'package:sp_settings/fields/button_settings_field.dart';
+import 'package:sp_settings/fields/popup_settings_field.dart';
 import 'package:sp_settings/settings_category.dart';
 import 'package:sp_settings/fields/settings_field.dart';
 import 'package:sp_settings/fields/switch_settings_field.dart';
+import 'package:sp_settings/settings_list.dart';
 
 void main() {
   runApp(const MyApp());
@@ -15,9 +17,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Settings Form',
-      theme: ThemeData(
-        primarySwatch: Colors.blue
-      ),
+      theme: ThemeData(primarySwatch: Colors.blue),
       home: const MyHomePage(title: 'Settings Form Demo'),
     );
   }
@@ -39,28 +39,63 @@ class _MyHomePageState extends State<MyHomePage> {
       appBar: AppBar(
         title: Text(widget.title),
       ),
-      body: const SettingsForm(
-        categories: [
-          SettingsCategory(title: 'General', fields: [
-            SwitchSettingsField(
-              settingsField: SettingsField(
-                title: 'Example',
-                description: 'This is an example.',
-                icon: Icons.add,
-              ),
-              prefKey: 'example1',
-            ),
-          ]),
-          SettingsCategory(title: 'Other', fields: [
-            SwitchSettingsField(
-              settingsField: SettingsField(
-                title: 'Example',
-                description: 'This is an example.',
-                icon: Icons.folder,
-              ),
-              prefKey: 'example2',
-            ),
-          ]),
+      body: ListView(
+        children: [
+          const SettingsCategory(
+              title: 'General',
+              settingsList: SettingsList(
+                fields: [
+                  SwitchSettingsField(
+                    settingsField: SettingsField(
+                      title: 'Example',
+                      description: 'This is an example.',
+                      icon: Icons.add,
+                    ),
+                    prefKey: 'example1',
+                  ),
+                ],
+              )),
+          SettingsCategory(
+              title: 'Other',
+              settingsList: SettingsList(
+                fields: [
+                  const SwitchSettingsField(
+                    settingsField: SettingsField(
+                      title: 'Long text switch',
+                      description:
+                          'This is an example. The description is a lot longer. It is very very long. It reaches three lines.',
+                      icon: Icons.folder,
+                    ),
+                    prefKey: 'example2',
+                  ),
+                  ButtonSettingsField(
+                    settingsField: const SettingsField(
+                      title: 'Clickable',
+                      description:
+                          'Here is a button setting that can be clicked.',
+                      icon: Icons.person_off,
+                    ),
+                    onTap: () {},
+                  ),
+                  const PopupSettingsField(
+                    settingsField: SettingsField(
+                      title: 'Pop-up',
+                      description:
+                          'Here is a pop-up menu! Select one of these.',
+                      icon: Icons.drive_eta,
+                    ),
+                    prefKey: 'popup',
+                    items: {
+                      'None': 'null',
+                      'Apple': 'a',
+                      'Banana': 'b',
+                      'Carrot': 'c',
+                      'Lemon': 'l',
+                      'Strawberry': 's',
+                    },
+                  ),
+                ],
+              ))
         ],
       ),
     );