+++ lib/fields/settings_field.dart
@@ -26,40 +26,45 @@ class _SettingsFieldState extends State<SettingsField> {
   @override
   Widget build(BuildContext context) {
     return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 8),
+      padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 16),
       child: Row(
         crossAxisAlignment: CrossAxisAlignment.center,
         children: [
           // 32px icon, centered
           if (widget.icon != null)
-            Icon(
-              widget.icon!,
-              size: 32,
+            Padding(
+              padding: const EdgeInsets.only(right: 16),
+              child: Icon(
+                widget.icon!,
+                size: 32,
+              ),
             )
           else
             Container(),
           // Title and description stacked vertically
           Expanded(
-            child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 8),
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Padding(
+                  padding: const EdgeInsets.symmetric(vertical: 4),
+                  child: Text(
                     widget.title,
                     overflow: TextOverflow.ellipsis,
+                    style: const TextStyle(fontWeight: FontWeight.w500),
                   ),
-                  if (widget.description != null)
-                    Text(
-                      widget.description!,
-                      overflow: TextOverflow.ellipsis,
-                      style: const TextStyle(fontWeight: FontWeight.w300),
-                    )
-                  else
-                    Container(),
-                ],
-              ),
+                ),
+                if (widget.description != null)
+                  Text(
+                    widget.description!,
+                    overflow: TextOverflow.ellipsis,
+                    maxLines: 3,
+                    style: const TextStyle(fontSize: 12),
+                  )
+                else
+                  Container(),
+              ],
             ),
           ),
         ],