+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.3
+
+* Replace DropdownSettingsField with PopupSettingsField.
+* Increase padding and size of category titles.
+* Introduce SettingsList as argument to SettingsCategory
+
 ## 1.0.2
 
 * Do not require prefKey for ButtonSettingsField.